* known-problems: Add G611-028 entry.
* vcs-generic_vcs.adb:
(Execute) [Parser_Command_Type]: Initialize the status record with the
 current cached one.

* vcs_module.adb:
(File_Edited_Cb): If the file status is unknown just ignore it as we
 do not want to override cache information.
(File_Status_Changed_Cb): If the file status is unmodified just return,
 there is nothing to do in this case. This is the hook run when a new
 file is opened.

* vcs_view-explorer.adb:
(Display_File_Status): Set the cache with the status if the current
 stored status is unknown.
--
The status on the VCS explorer were reset each time a file was opened.
For G611-028.



git-svn-id: svn+ssh://svn.eu/Dev/trunk/gps@108247 936e1b1b-40f2-da11-902a-00137254ae57