Highlight entry in revision view if already present.
(Add_Log_If_Not_Present): If a node is already present in the
 revision view we want to highlight.

For example after opening the revision view for all revisions
and adding annotations to the file, clicking on any revision
on the annotated file must highlight the corresponding entry in
the revision view instead of staying silent.
--
Work for H103-004.


git-svn-id: svn+ssh://svn.eu/Dev/trunk/gps@120505 936e1b1b-40f2-da11-902a-00137254ae57