(ExpSlaveDebugThread): Do not add the main process in the wait queue. (ExpCommonDebugger): Remove obsolete and inacurate comment. (OnXCreateProcess): Do not add child processes in the wait queue.
This modification makes processes (debugged one and its subprocess) not being
added to the main explaunch wait queue, as this is not needed - the main wait
procedure waits for the debug thread to terminate, and this thread terminates
only when all processes are finished - and this caused problems on Vista where
not all processes terminations are signaled.

Minor reformatting, remove some #if 0.

Correction of H108-008


git-svn-id: svn+ssh://svn.eu/Dev/trunk/gps@120516 936e1b1b-40f2-da11-902a-00137254ae57