Smarter activity menu for adding/removing files.
In the VCS contextual menu the submenu to add a file
to an activity was always present even if not working
when the file was already part of an activity. Now the
submenu "remove from activity" is proposed instead of
the add one in this case. Use the pago emphasize markup
for cleaner menu (and to match the current design for
other menu entries).

* vcs/src/vcs_activities_view_api.adb:
  (On_Menu_Remove_From_Activity): Procedure spec moved from here
   to spec. Add support for file context.

* vcs/src/vcs_activities_view_api.ads:
  (On_Menu_Remove_From_Activity): Procedure moved here
   from body.

* vcs/src/vcs_view_api.adb:
  (VCS_Contextual_Menu): Display either the add or remove
   file to/from an activity submenu.

* docs/gps.texi:
  Add corresponding documentation.

* distrib/features:
  Add corresponding entry.
--
Work for H111-019.


git-svn-id: svn+ssh://svn.eu/Dev/trunk/gps@120723 936e1b1b-40f2-da11-902a-00137254ae57