Add revision numbers on head of the list.
Always add the new revisions (working/repository) on the head
of the list. If the cache already contains revision information
we want the new ones to be the first on the list as those are
the revision information displayed on the VCS Explorer and
associated with the corresponding files.
--
Fixes H309-006.


git-svn-id: svn+ssh://svn.eu/Dev/trunk/gps@122930 936e1b1b-40f2-da11-902a-00137254ae57