Fix adding single files in an activity.
Adding a single file from a source editor was not working properly
as the context information was wrongly updated. When building the
VCS contextual menu we want to keep the creator of the context. This
is used later to know where to look for the selected files.
File a non regression test.
--
Fixes H314-018.


git-svn-id: svn+ssh://svn.eu/Dev/trunk/gps@123050 936e1b1b-40f2-da11-902a-00137254ae57