Fix removing multiple files from the VCS explorer.
Problem uncovered while working on H314-018. It was not
possible to remove from an activity multiple files selected
in the VCS Explorer. A multiple file selection on the
On_Menu_Remove_From_Activity was always taken from the
Activities Explorer. We now check the widget with the focus
to either use the VCS Explorer or VCS Activity Explorer
selection. At the same time we properly check for the
module creator for the single file selection with context
coming from a source editor.
--
Fixes H316-002.


git-svn-id: svn+ssh://svn.eu/Dev/trunk/gps@123052 936e1b1b-40f2-da11-902a-00137254ae57