Add a Context_Factory for the VCS Activities.
The VCS had not context factory. This part was conplely broken.
First the callback in VCS Module had not the proper spec and
only the VCS explorer context was called. The former problem
has already been solved in a previous commit. This new change
set add the proper context factory for the VCS Activities
explorer and ensure that it is called when this child is
selected.
--
Continued work for H405-002.


git-svn-id: svn+ssh://svn.eu/Dev/trunk/gps@123782 936e1b1b-40f2-da11-902a-00137254ae57