Store a File_Array instead of a single file into the context.
This change make it possible to handle properly the case of the
VCS Explorer and the VCS Activities Explorer with multiple files
selected. It was previously needed to check the current selected
child and get the corresponding selected files. This was not
possible to keep simple the implementation to support all cases.
--
This patch makes the VCS Activities action support cleaner.
It make also possible to clean the trick made to build the
VCS dynamic contextual menu (no more modification of the
context while building the menu).
--
Add a corresponding test case.
--
Work for H405-002 and H403-008.


git-svn-id: svn+ssh://svn.eu/Dev/trunk/gps@123783 936e1b1b-40f2-da11-902a-00137254ae57