Fix get specific revision action.
Because of the way the file are handled with Subversion
it was not possible to get a specific revision of a file.
We now use the new base_name and set_title actions to properly
separate the filename and the buffer title. This is cleaner and
make it possible to have a cleaner title for buffer containing
a file specific revision.
--
Part of work for H504-003.


git-svn-id: svn+ssh://svn.eu/Dev/trunk/gps@124619 936e1b1b-40f2-da11-902a-00137254ae57