Fix diff2 Subversion action, improve CVS diff2.
This action was not properly implemented and has never
worked. As both revisions can be different to the head
one (i.e. the one on the edited buffer) we need to get
the specific revisions from the repository first.

Also use a better editor name for the CVS diff2 action.
This matches what is done on Subversion and will ease
writting automatic tests.
--
Last part of work for H504-003.


git-svn-id: svn+ssh://svn.eu/Dev/trunk/gps@124620 936e1b1b-40f2-da11-902a-00137254ae57