Fix read-only file support.
Do not use the file descriptor as part of the check
for read-only files. This is boggus as Write_File write
the content into a temporary file which is never read
only. The previous check prevented the dialog asking
for overwrite to be displayed. Also in the case of
overwrite confirmed by the user do not call Write_File
as the content as already been written into the temporary
file.
--
Fixes H602-007.


git-svn-id: svn+ssh://svn.eu/Dev/trunk/gps@125435 936e1b1b-40f2-da11-902a-00137254ae57