* check_button.c: (gtkada_check_button_install_handler): Allow the complete replacement of the  default handler for 'clicked' as this handler is always called first and we  need to be called before it. (gtkada_check_button_force_state): Directly modify the check button active  state so that the call to the clicked handler behave as we want.
* gtkada-check_button.adb:
(On_Expose): Modify the profile so that it can be directly called by gtk.
(Initialize): Create a new widget class, and replace the default 'clicked'
 class handler.
(State_Changed): Remove this now obsolete method.
(Set_Default): Better handling of the signals.
(Set_Active): Do not send signals if the state did not change.
(Clicked): Change the state so that the final state as seen after the
 On_Clicked handler is run is the initial state
(On_Clicked): Better handling of the underlying gtkcheckbutton state. This
 prevents to send multiple signals for a single action.

* gtkada-check_button.ads:
(Clicked): Override the parent method so that the internal state is correctly
 handled.

* switches_chooser.adb:
(Change_Switch_Status): Moved outside of Change_Switch and renamed as
 Handle_Dependencies. This is now shared with
 On_Command_Line_Change. Correction of the Remove_Switch call, and exit the
 loop as soon as the slave switch is found.


Continuing corrections for H605-014




git-svn-id: svn+ssh://svn.eu/Dev/trunk/gps@125618 936e1b1b-40f2-da11-902a-00137254ae57