Fix the activity contextual menu with multiple selected files.
When selecting multiple files on the VCS Explorer the
activity oriented entries in the contextual menu where not
always accurate. The menu "Commit as new Activity" or "Add to
Activity" could be present even though a selected file was
already in an activity. The menu "Remove from Activity" was
not listed if the first selected file had no activity.

It was easy to fix those issues now that the context contains
all the selected files. File a corresponding regression test.
--
Fixes H409-028.

git-svn-id: svn+ssh://svn.eu/Dev/trunk/gps@125836 936e1b1b-40f2-da11-902a-00137254ae57