Use a meaningful title for the text revision window.
This is especially important for the Subversion VCS as the
name of the file is not output by the external command. The
name is output by cvs VCS. It is anyway cleaner to use a
proper title for such windows instead of using the GNAT
temporary file template.
--
Fixes H904-019.

git-svn-id: svn+ssh://svn.eu/Dev/trunk/gps@128374 936e1b1b-40f2-da11-902a-00137254ae57