* html.tmpl: Use <pre> tags for code printout, instead of white-space: pre style, the latter  being incompatible with IE6
* docgen.css:
Remove the default margins of the <pre> tag.
Do not use the white-space style, as this is not compatible with IE6

* docgen2.adb:
(Generate_Annotated_Source): Use the new Docgen2_Backend.Handle_Code methods
 to generate annotated source with line feeds compatible with more browsers.
 This also allows line numbering now.

* docgen2_backend-html.adb:
(*_Handle_Code): New. Allow handling of code output, with better handling of
 line feeds and line numbering.

* docgen2_backend-html.ads:
(*_Handle_Code): New. Allow handling of code output, with better handling of
 line feeds and line numbering.

* docgen2_backend.ads:
(*_Handle_Code): New. Allow handling of code output, with better handling of
 line feeds and line numbering.

Remove use of "white-space: pre" style, as this is not compatible with IE6.




git-svn-id: svn+ssh://svn.eu/Dev/trunk/gps@128500 936e1b1b-40f2-da11-902a-00137254ae57