* build_command_manager.adb: (Launch_Target): Refine Quiet mode. Implement Command API.
* build_command_manager.ads:
Add command API.

* builder_facility_module.adb:
Add support for "On file save" target launch mode.
(Customize): Factorize code with Attempt_Target_Register.


* builder_facility_module.ads:
Update documentation.

* commands-builder.adb:
Send each compiler output to a different category in the Locations View.
Improve support of Quiet mode.
This is needed for the support of "On file save" modes.

* commands-builder.ads:
Send each compiler output to a different category in the Locations View.
Improve support of Quiet mode.
This is needed for the support of "On file save" modes.

G706-005.




git-svn-id: svn+ssh://svn.eu/Dev/trunk/gps@129089 936e1b1b-40f2-da11-902a-00137254ae57