* dualcompilation.adb: (Set_Dualcompilation_Properties): Save as much as possible, even when active  is not set. (Locate_Exec): New, only searches in the selected path. (Internal_Locate_Exec): Make use of Locate_Exec.
* dualcompilation.ads:
(Locate_Exec): Make this method accessible, so that we can check the user paths
 before closing the dual compilation preference dialog.

* dualcompilation_dialog.adb:
Add the browse buttons to allow easy selection of the paths, and protect the
 callbacks against exceptions.
Add tooltips on the text entries to explain what the two paths mean.

* dualcompilation_dialog.ads:
(Gtk_New): Add the kernel as parameter, and remove Parent now not needed.

* dualcompilation_module.adb:
(On_Menu): Use actual compiler used by the project to determine the default
 paths, and also determine if the user-selected paths are correct. Display
 a dialog if the selected compiler path does not contain any compiler.

Complete the implementation of the dual compilation mode.

Part of H229-002




git-svn-id: svn+ssh://svn.eu/Dev/trunk/gps@129210 936e1b1b-40f2-da11-902a-00137254ae57