Ensure that location row is visible on the location view.
This is a rewrite of the On_Row_Expanded routine. We now use
an idle callback to be sure that the tree view is properly
updated. At this point it is possible to get the visible
paths (start and end). From this we make sure that the
child of the expanded row is visible on screen.
--
Fixes HA16-020.

git-svn-id: svn+ssh://svn.eu/Dev/trunk/gps@131050 936e1b1b-40f2-da11-902a-00137254ae57