Merged from HEAD: When selecting a file set the file dialog to "open" mode.
The mode was not specified (and defaulted to Unspecified). In this
case using a browse button to select a file in the project properties
a dialog was opened with the button named "save" instead of "open".

This was causing a problem on Windows Vista when the file was read-only
for example.

Fixes HA14-009.

git-svn-id: svn+ssh://svn.eu/Dev/branches/global/gps-4_3-branch/gps@131489 936e1b1b-40f2-da11-902a-00137254ae57