Add support for automatic selection of the VCS.
It is now possible to use "auto" for the project's VCS_Kind
attribute value. In this case GPS will try to guess the proper
VCS to use. This is done by looking for the administrative
VCS directory (.svn, CVS, .git...). The new generic VCS
administrative_directory attribute can be used to specified
this information.

As returning the VCS for a project can now be costly this information
has been cached using an hash map.

Note also that the VCS registration circuitry has completely
been rewriten to something simpler and faster. All VCS loaded
(or native) is now registered into a map the key being the VCS
name. All this central VCS registry has been moved to the
VCS_Module package. This also fixes a memory leak, now when
the module is destroyed we properly free all VCS information.
Because of this work the specific code for destroying the
Clearcase module has been removed.

A corresponding test has been added.

Work for DB04-011.

git-svn-id: svn+ssh://svn.eu/Dev/trunk/gps@131554 936e1b1b-40f2-da11-902a-00137254ae57