Add new preference No-Implicit-Status.
This new boolean preference controls whether a status command
can be launched automatically after some other actions. If the
network connection is slow it is better to set this value to
true to avoid such status actions. This is for example done
after an update to retrieve the actual revision number of the
new files. So, when set to True some revision number may not
be accurate. The solution is to run a status action explicitly.
--
Work for H927-002.

git-svn-id: svn+ssh://svn.eu/Dev/trunk/gps@131682 936e1b1b-40f2-da11-902a-00137254ae57