Add new action update_parse.
This new action is used to parse the output of the update
action and to update the file statuses accordingly. This
will make it possible to have somewhat accurate file
status even if no status action is run after an update.

Note that with CVS we need the update_parse action in
the on-failure not too as CVS return with an error if
there is merge conflics for example.
--
Work for H927-002.

git-svn-id: svn+ssh://svn.eu/Dev/trunk/gps@131683 936e1b1b-40f2-da11-902a-00137254ae57