Merge revision 131479 for HA29-013:
Add support for non ASCII characters on GPS root dir.

The .gps directory is created under HOME or GPS_HOME directory.
If this directory contains non ASCII characters on Windows we
need to convert them to Utf-8 (this is the expected encoding
by the GNAT runtime). At the same time we ensure that every
external names passed to the Windows OS are UTF-16 encoded.
--
Fixes HA05-007.


git-svn-id: svn+ssh://svn.eu/Dev/branches/global/gps-4_3-branch/gps@131783 936e1b1b-40f2-da11-902a-00137254ae57