Add way to select VCS_Kind from an external variable.
On UNIX it was set to Auto and Subversion on Windows
to avoid confusion between the two Subversion (Native, Cygwin)
implementations. It is now possible to select the VCS to use
from and external variable on Windows. The default being set
to Subversion. This is not needed on UNIX which is set to Auto.
--
Fixes HB02-011.

git-svn-id: svn+ssh://svn.eu/Dev/trunk/gps@131878 936e1b1b-40f2-da11-902a-00137254ae57