Move action labels support in root class VCS.
The labels for every VCS actions is used by the generic VCS
and the clearcase native. Both had a copy of the labels. It
is cleaner to share this information and avoids having both
to redefine Get_Identified_Actions routine. This routine
is now promotted to class-wide.
--
Clean-up for DA05-003.

git-svn-id: svn+ssh://svn.eu/Dev/trunk/gps@131894 936e1b1b-40f2-da11-902a-00137254ae57