Filter VCS menu to display only available actions.
The main VCS menu is filtered to contain only the VCS actions
that are at least defined once in a generic VCS implementation.
To achieve this, the main VCS menu constructor is deferred
after GPS is started as we must build it after all VCS plug-ins
are loaded.
--
Work for DA05-003.

git-svn-id: svn+ssh://svn.eu/Dev/trunk/gps@131898 936e1b1b-40f2-da11-902a-00137254ae57