Use Filename_[To/From]_UTF8 instead of Locale_[To/From]_UTF8.
This is a temporary work. Filenames should be translated for
display purpose only (the corresponding servives in GNATCOLL.VFS
should be used). This first step will help finding all places
where a filename conversion occurs.
Note that Filename_[To/From]_UTF8 are no-op routine on Windows.
--
Continued work for HA29-013.

git-svn-id: svn+ssh://svn.eu/Dev/trunk/gps@132086 936e1b1b-40f2-da11-902a-00137254ae57