Use Filename_To_UTF8 instead of Locale_To_UTF8 for menu entries.
The Initialize routine is used to add items into the menu containing
the recent project files. More generally this routine deals with
Full_Path_Menu_Item, we want to use the Filename_To_UTF8 as this does
nothing on Windows (the filenames are reported in UTF-8 by the GNAT
runtime).
--
Fixes HB05-026.

git-svn-id: svn+ssh://svn.eu/Dev/trunk/gps@132087 936e1b1b-40f2-da11-902a-00137254ae57