Add a way to enable/disable Clearcase module.
The clearcase module cannot be easily disabled by users. The fact
that it has been reverted to activated by default has broken the
fix for DA05-003.

This change is also important for DA05-006 (remove notion of log)
as we really want to remove the log oriented VCS menus
(like Edit revision log) when no VCS require work. So this is also
a preparatory work for DA05-006.

git-svn-id: svn+ssh://svn.eu/Dev/trunk/gps@135108 936e1b1b-40f2-da11-902a-00137254ae57