Add support for require_log generic VCS attribute.
It is now possible to create a VCS that do not require a log
(logs are handled by external commands like ClearCase GUI for
example). In this case GPS won't ask for a log file while
doing the commit/add/remove VCS actions.
The log oriented contextual menu are also removed in this case.
--
Work for DA05-006.

git-svn-id: svn+ssh://svn.eu/Dev/trunk/gps@135111 936e1b1b-40f2-da11-902a-00137254ae57