Select_File returns un-translated Virtual_File.
This routine used to return the name encoded in UTF-8. This was
wrong in many places. It is better to keep the filename as-is and
translate it when needed for display purpose using Display_Full_Name
or Diplay_Base_Name. This change has also avoided one convertion
from UTF-8.
--
Continued clean-up for HB05-017.

git-svn-id: svn+ssh://svn.eu/Dev/trunk/gps@135756 936e1b1b-40f2-da11-902a-00137254ae57