Fail to commit activities on Windows
GPS could fail to commit activities using Cygwin/Subversion
on Windows when a filename is passed on the command line with
the wrong casing. Windows is non case-sensitive but keeps
filename casing. Using different casing for the pathname
confuses Cygwin/Subversion which checks that all files, part
of a group commit, are all under the same repository.
Add corresponding regression test.
--
Fixes I117-011.

git-svn-id: svn+ssh://svn.eu/Dev/trunk/gps@135916 936e1b1b-40f2-da11-902a-00137254ae57