Add custom sort on the revision view.
To have proper sorting on the date and revision number custom sort
routines have been introduced. The date sort routine is checking
for expanded date format (Git date format). Other VCS are outputing
date which can be sorted by a regular string comparison. For the
revision number the CVS dotted notation is checked, when revision
are plain number (like on Subversion) a numeric order is used.
--
Work for I304-009.

git-svn-id: svn+ssh://svn.eu/Dev/trunk/gps@141083 936e1b1b-40f2-da11-902a-00137254ae57