Do not escape backslahes in pathname substitution.
This is important as on windows an UNC path \\server\share would
be converted to \\\\server\\share. Such pathes won't be detected
as valid UNC path by the Win32 API. Note that this was not dected
before as it is fine to double directory separators in non remote
path (e.g. c:\\dir\\subdir is ok for Win32).

As a consequence, this patch clean the output message sent into the
GPS console where no duplicate directory separators are output on
Windows plateform.
--
Fixes I326-017.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@141964 936e1b1b-40f2-da11-902a-00137254ae57