Fix update of the read/write status.
The read/write status was not properly updated in all cases
(at least on Windows). The Check_Writable() call in the
idle procedure has been moved at the start of the routine
as we want to check for the read/write status in all cases.

Note that this routine has complex circuitry to avoid recursivity
GPS when trying to open the reload dialog. This case does not
apply here as no dialog is to be opened, so it is safe to move
the check outside of this section.
--
Fixes I414-019.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@143273 936e1b1b-40f2-da11-902a-00137254ae57