Ensure that VCS actions are based on reference file.
For auxiliary buffers we now set in the context the reference
file. For example, from a diff file the contextual menu action
will have the actual file to be the reference file. This ensure
that for example the commit will be recorded for the version
controlled files and not the temporary file used to display the
diff. It is then possible to create the revision log or ask for
annotations from an auxiliary buffer.

To achieve this the new VCS.set_reference (commited previously) is
used when creating diff buffers.

Add corresponding regression test.
--
Work for I425-006.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@143450 936e1b1b-40f2-da11-902a-00137254ae57