Do not replace with identical text.
This is an optimization of the on-the-fly casing mode. In the
case where the text to be replaced is identical to the replacement
we do not perform the replacement. This saves a replacement and
an undo recording. It also fixes some unexpected cursor movement
when deleting some text.
--
Work for I511-016 and fixes I512-018.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@143963 936e1b1b-40f2-da11-902a-00137254ae57