Do not move backward in On_Th_Fly casing.
Moving backward is only needed for End_Of_Word casing as we
want to move back to the word just endded. It is especially
important to not move backward when deleting a character as
we could move out of a string literal (deleting first character
on a string) and change its casing.

This ensure that in:

   C : constant String := "azerty";

If deleting the character 'a', no casing modification is done
inside the string.
--
Fixes I609-003.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@145410 936e1b1b-40f2-da11-902a-00137254ae57