Implement On-The-Fly conservative casing.
This is used in the On-The-Fly casing and ensures that the casing
occurs for the characters as typed by the user. For example if
the user enter characters "NEWT" GPS will recase this way:

  Typed       Buffer
  -----       ------
  N           N
  E           NE
  W           new
  T           NEWT

(when keyword casing is set to lower)

Previously the buffer would have ended as "NewT".

Add corresponding regression test.
--
Work for I616-051.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@146867 936e1b1b-40f2-da11-902a-00137254ae57