Merged from trunk: Refresh the activities explorer when a project is loaded.
When a new project is loaded, or after loading the initial
project we want to refresh the activities explorer. This is
needed as only at this point the root project is known and
this information is needed to display the corresponding
activities.

Add corresponding regression test.
--
Fix regression reported on I930-007.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/branches/global/gps-4.4/gps@150211 936e1b1b-40f2-da11-902a-00137254ae57