Fix language detection regular expression.
On gdb 6.8 and 7.0 the language is displayed either with:

   Current language:  auto; currently c
   The current source language is "auto; currently ada".

With gdb 7.0 we have:

   Current language:  auto

Which just remind the current language detection mode. We want to
skip this line which does not contain language information.

In the new regular expression the "auto; currently" string is not
mandatory anymore.
--
Fix regression on debug.parsing.1.
Work for IA12-004.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@151329 936e1b1b-40f2-da11-902a-00137254ae57