Makes the call tree view shared.
Instead of opening a new windows for every call tree we now
reusing the window. To achieve that the kind of tree (calls
or called by) is now stored in the call tree and not a
view attribute anymore.

Update IA27-021 and H131-008 tests accordingly.

This makes the Tools -> Views -> Call Tree menu useful.
--
Work for IC07-012.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@154018 936e1b1b-40f2-da11-902a-00137254ae57