Filters can match multiple times.
(Text_Output_Filter): Make sure that a filter is used on all
possible matches. This has been done to be able to properly match
an output with:

   Current language:  auto
   The current source language is "auto; currently c".

Previously we were matching only the first line and if the
current selected language was Ada we were not properly swithing
to C language specified in the second line.
--
Fixes J131-002.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@156051 936e1b1b-40f2-da11-902a-00137254ae57