* basic_types.ads: (String_Index_Type): New type, to identify clearly when we are dealing with  string indexes.
* entities-queries.adb:
(Find_Declaration): Reactivated the use of heuristics in case in fuzzy match
 now that problems with refactoring are sorted out.
 Use the ALI entity if found at the place detected by the construct heuristics.

* language-tree-database.adb:
(To_Line_Column): New subprogram.

* refactoring-parameters.adb:
(Is_Dotted_Notation): Used a more accurate version, taking advantage of the
 reverse expression parser to get the actual package reference, instead of
 heuristics leading to fuzzy matches.

* entities-tooltips.adb:
(Get_Documentation): Fixed retreival of the line offset.

This patch introduces a new type, String_Index_Type, used when we need to
 clearly idendify that we're dealing with string indexes (starts at 1, and
 expressed in byte offsets), and started to use it in the language construct
 machinery.
Completes and refine the implementation of heuristics for navigation.
Part of J208-007.




git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@156773 936e1b1b-40f2-da11-902a-00137254ae57