* ada_semantic_tree-lang.ads: (Ada_Handler): New type.
* docgen2.adb:
(Execute): Now manages a lock on the construct database so that we avoid
 computing the thing too many times.

* entities.ads:
(Abstract_Language_Handler_Record): Moved to languague.tree.database.

* language-tree-database.ads:
(Abstract_Language_Handler_Record): Moved from languague.tree.database.
(Get_Or_Create): Now deduces the language directly from the database.

This patch moves Language_Handler to Language.Tree.Database, so that is't
 possible to associate the handler directly at the construct database level and
 we don't need to give languages to Get_Or_Create anymore.
It also add locks while computing the documentation so that the process is
 speed up.
Part of J208-007.




git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@156811 936e1b1b-40f2-da11-902a-00137254ae57