Set Save_Complete to False for new file.
A file just opened that does not exist on disk has never
been saved, so we initialize Save_Complete to False in this
case. Moreover, this is needed to know if a file that is not
on-disk has neved been saved or was removed outside GPS.
--
Preparatory work for J422-030.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@159527 936e1b1b-40f2-da11-902a-00137254ae57