Reactivate reading of timestamp of file before updating the contents database.
The timestamp of files on the filesystem is only set with second precision, and
the testsuite might do the following: write the file, fill the database, then
write the file again, then fill the database again, all within the same second.
This does not correspond to real-time use: introduce a delay of 1.0 second between
consecutive writes so that the timestamp is modified between one write and the
next.

For J503-017.


git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@159814 936e1b1b-40f2-da11-902a-00137254ae57