* gps.texi: Add documentation for new preference.
* gps-kernel-preferences.adb:
(Register_Global_Preferences): Register preference Current_Thin_Line.

* gps-kernel-preferences.ads:
(Current_Line_Thin): New preference.

* src_editor_view.adb:
(Change_Handler): Invalidate the whole window, this is needed for "thin line"
 highlighting of current line and upcoming background compilation.
(Highlight_Text): Draw current line as a thin line if the preference is set
 correspondingly.
(Execute): Store the value of the preference that controls the aspect of the
 current line.

For J621-010.




git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@161390 936e1b1b-40f2-da11-902a-00137254ae57