* features: Add entry for J813-008.
* gps.texi:
Minor adaptation: in the Fonts & Colors section of the preferences, we are now
 selecting font variants, rather than fonts.

* default_preferences.adb:
Add support for Variant_Preferences.

* default_preferences.ads:
Introduce a type "Variant_Enum" to control the font variants supported in GPS.
Add a new type of preference: Variant_Preferences.
Add supporting functions.

* gps-kernel-preferences.adb:
(Register_Global_Preferences): Create the preferences for fonts in the editor
 as Variant preferences, not style preferences.

* gps-kernel-preferences.ads:
Make all preferences for the source editor fonts Variant preferences, except
 for the default font.

For J813-008.




git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@164194 936e1b1b-40f2-da11-902a-00137254ae57