* toolchains.adb: (Set_Compiler): Handle the case where we set a user-defined compiler and no previous compiler was defined before. (Clear_Toolchains): Also clear the new saved toolchains list (Do_Snapshot, Do_Commit, Do_Rollback): New, for handling cancellation of toolchains edition. (Free): New.
* toolchains.ads:
Implement snapshot/rollback/commit for a manager to be able to handle
cancelation of toolchains edition.

* toolchains_editor.adb:
Cache the manager in the module, so that a full gprconfig scan can be done
only once during a GPS session.
(Set_Project): Use rollback and snapshot so that we handle clicking on Cancel.
(On_Server_Changed): Completely free the toolchains manager, totally useless as
soon as a new Build_Server is selected.

Merge of JA04-027 and J924-018




git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/branches/global/ide-2010/gps@165381 936e1b1b-40f2-da11-902a-00137254ae57