* builder_facility_module.adb: (Append_To_Build_Output): Append to the output without modifying the container,  saves a big amount of time in adjust/finalizes.
* builder_facility_module-scripts.adb:
(Shell_Handler): Add support for new parameter "as_string" to
 "get_build_output".
(Register_Commands): Adjust number of arguments to "get_build_output".

* codefix_module.adb:
(Compilation_Finished_Cb): Take advantage of new parameter "as_string" to
 command "get_build_output".

* shell_commands.xml:
Add documentation for new parameter "as_string" to get_build_output.

The command "get_build_output" is returning the output as a list by default.
When this is done on the 'Shell' scripting language, the implementation splits
 the string at every ASCII.LF, only to recreate it in Set_Return_Value.

This is a useless performance bottleneck when manipulating a huge build output.

To circumvent this, this change introduces a new parameter "as_string" that
 asks for the output to be returned as a raw string, and make sure that
 codefix uses this.

For JB01-014.



git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/branches/global/ide-2010/gps@166891 936e1b1b-40f2-da11-902a-00137254ae57