Fix key manager conflict detection.
The key manager recording was too restrictive. For example it was not
properly handling the case where we define a new mapping for ctrl-x when
some other actions are mapped to key using ctrl-x as prefix (e.g. ctrl-x+b).
In this case we now display all current actions using this prefix and
ask user if it is ok to delete all those and create the new binding.

Another limitation was that it refuses to define a binding for ctrl-x+b if
some actions were using ctrl-x as prefix.

Work for JB07-002.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@169330 936e1b1b-40f2-da11-902a-00137254ae57