Create new files in user friendly location.
GPS now proposes to store new files in a more user friendly way. The
default directory to create the file depends on the context. So for
example if a specific directory is selected into the project explorer
the File->New action will record this. This specific location
will be proposed by default when doing File->Save.

To achieve that, we now store the directory information into the
Virtual_File associated with a buffer.

* src_editor_box.adb:
  (Save_To_File): Pass base directory to Select_File when saving.

* src_editor_module.adb:
  (Create_File_Editor): Handle Is_Directory the same way No_File is
   handled.
  (Open_File): Likewise.
  (On_New_File): Use directrory information found in context if any.

Add corresponding feature entry.

For K518-001.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@174016 936e1b1b-40f2-da11-902a-00137254ae57