Implemented MI.Utils.Process_Var_Create.
This commit adds the implementation of the Process_Var_Create subprogram to the
already implemented set of subprograms declared in this package.
It also adds a test case for this routine. This test case introduce the
Read_File_Helper package defined in the test sources that implements a set of
helper functions to easily extract the result of the MI parser, given a file
containing MI output. This allows easy crafting of AST node for better testing.

src/ChangeLog:

        * mi-utils.adb (Process_Var_Create): Removed stub default
          implementation and replaced it by its final implementation.

tests/src/utils/ChangeLog:

        * read_file_helper.ad{b,s}: New files that declares and implement
        helper subprograms for testing purpose.
        * process_var_create_test.adb: New test case for the Process_Var_Create
        routine.

tests/utils/ChangeLog:

        * var-create-test-1.mi: Content of the Process_Var_Create test case.

Part of [K526-005].

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@175535 936e1b1b-40f2-da11-902a-00137254ae57