* gps-kernel-messages.adb: (On_File_Renamed_Hook): New hook, to react to File_Renamed. (Create_Messages_Container): Register it.
* src_editor_buffer.adb:
(Save_To_File): Emit "File_Renamed" if we are renaming the file.

Found thanks to L124-025:

When we rename an editor, emit the File_Renamed hook, and in the Messages
handling, react to this hook and remove all messages belonging to the editor.
This avoids potential dangling references to messages being held by the
source editors after a rename.




git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@185809 936e1b1b-40f2-da11-902a-00137254ae57