Fix issue with commit command in the Mercurial VCS plugin.
The command is given a file as commit message. If this file is empty, hg
spawn the editor, which blocks GPS. The fix it to add a trailing
whitespace in the message file that simulate an empty commit message.

This trailing whitespace is removed by hg during the commit.

L405-002.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@192032 936e1b1b-40f2-da11-902a-00137254ae57