------------------------------------------------------------------------ r199203 | setton | 2012-12-05 17:50:13 +0100 (Wed, 05 Dec 2012) | 7 lines
Make the "color depth check" dialog depend on a trace
CHECK_COLOR_DEPTH rather than on the trace TESTSUITE
which is too general.
Reactivate tests that were blocking on this dialog.
This is needed because the display depth is not any
more stable on Windows than it is on Xvfb.

------------------------------------------------------------------------


git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/branches/global/gtk3/gps@199204 936e1b1b-40f2-da11-902a-00137254ae57