* kernel/src_info/old_entities.ads (Reference_Kind): Adding "unknown" value. (Entity_Reference_Index): Adding "kind" component to avoid searching  for its value in the ordered set of references if already known. (Null_Entity_Reference): Adding value of new component.
* kernel/src_info/old_entities.adb
(Lt_No_File): Allow both 'r' and 'm' references at the same location.
(Index): Return "unknown" as kind of reference
(Get_Kind): If the kind of the entity is already known then return it
 immediately; otherwise search for the reference in the ordered set of
 references to retrieve its value (legacy functionality).
(Kind_To_String): Adding case "Unknown"

* kernel/src_info/old_entities-queries.adb
(Get): Return the kind of reference available through the iterator.

For M103-017

[GPS] Write references shown as read references

Several routines of package Old_Entities assume that there is always a
single reference associated with each location. However, since 2010-11-05
the compiler generates two references associated with OUT and INOUT
parameters (see revision 170013 in the GNAT Repository)

This patch workarounds the problem as follows: when iterating through
entity references, the kind of the entity is returned in the corresponding
entity reference index (which is later used to retrieve the reference and
also its kind).


git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@200658 936e1b1b-40f2-da11-902a-00137254ae57