("Display type hierarchy"): contextual menu only displayed on types, not vars.
In the sqlite backend, a type does not return any value for the Type_Of query.
Instead, it returns values for Parent_Types. Since the context menu is
really meaningful on types anyway, the implementation was changed to test
for Parent_Types, and removed when clicking on variables.
Fixes F819_004_display_type_hierarchy.
M212_014


git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/branches/global/gtk3/gps@202641 936e1b1b-40f2-da11-902a-00137254ae57