* language-tree.ads: (Categories_For_Block_Highlighting): Add a constant declaration.
* src_editor_buffer.adb:
(Forward_Position): Do not do anything if the initial position is not valid.

* src_editor_module-editors.adb:
(Get_Block): Make use of Categories_For_Block_Highlighting.
 Address ??? comment.
(Create_Editor_Location): Add checks to verify that the position is valid
 before creating it.

* src_editor_view.adb:
(Change_Handler): The current block should be gotten from the list of
 categories to consider for block highlighting.

Follow up on I410-013.




git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/branches/global/gtk3/gps@206486 936e1b1b-40f2-da11-902a-00137254ae57